Add file watcher
Add /watch route (renamed from /wait) that will watch the .git/logs
folder for changes and return a json list of them.  Modified graph.js to
watch for changes and use a new route /svg to reload the graph.  Also
removed the poor template functionality in favor of dom manipulation on
the client side.