Add watch route
Add watch route that waits for changes to the git repo.  Added file
watchers on both the .git/logs and .git/refs folders that emit events on
a custom event emitter.  Created 55 second interval that emits a
heartbeat event that will prevent the browser from timing out; the
browser will accept the heartbeat and reissue the request for changes.

Created the start of a promise based non-blocking sequential function to
generate the dot language.