Add refs to graph
Add ref nodes to the graph with labels.  Included HEAD and indicated if
it is detached or not.

Added md5 hash to produce consistant colors for branches with a better
randomizer.  Ensured that master is redish.  The old hash is discarded
as it was biased.

Fixed queue_cmd function to only catch a promise error if requested.
Otherwise it was silently catching all errors.