Add option to include forward branches
The default branch list if unspecified will be just `master`.  Added the
option to look for all branches that include the last commit of master
and add them to the list by default.  This can be turned off my setting
`include_forward` to false in the settings object.

This will produce a short graph of changes not in master yet as the
default view.

Added a series of comments at the top of the digraph listing the git
commands issued to build the graph.

Removed the test call and added a modules.export.